Configure redux to save router state and configure redux dev tools.
