--- manoseimas/compatibility_test/client/app/main.jsx
@@ -1,24 +1,0 @@
-import React, { PropTypes } from 'react'
-import { Router } from 'react-router'
-import { Provider } from 'react-redux'
-
-class AppContainer extends React.Component {
-
-  static propTypes = {
-    history: PropTypes.object.isRequired,
-    routes: PropTypes.object.isRequired,
-    store: PropTypes.object.isRequired
-  }
-
-  render () {
-    const { history, routes, store } = this.props
-
-    return (
-      <Provider store={store}>
-        <Router history={history} children={routes} />
-      </Provider>
-    )
-  }
-}
-
-export default AppContainer
