+++ manoseimas/compatibility_test/client/store/configureStore.js
@@ -1,0 +1,44 @@
+import { applyMiddleware, compose, createStore } from 'redux'
+import { routerMiddleware } from 'react-router-redux'
+import thunk from 'redux-thunk'
+import makeRootReducer from './reducers'
+
+export default function configureStore (initialState = {}, history) {
+  // ======================================================
+  // Middleware Configuration
+  // ======================================================
+  const middleware = [thunk, routerMiddleware(history)]
+
+  // ======================================================
+  // Store Enhancers
+  // ======================================================
+  const enhancers = []
+  // if (__DEV__) {
+    const devToolsExtension = window.devToolsExtension
+    if (typeof devToolsExtension === 'function') {
+      enhancers.push(devToolsExtension())
+    }
+  // }
+
+  // ======================================================
+  // Store Instantiation and HMR Setup
+  // ======================================================
+  const store = createStore(
+    makeRootReducer(),
+    initialState,
+    compose(
+      applyMiddleware(...middleware),
+      ...enhancers
+    )
+  )
+  store.asyncReducers = {}
+
+  if (module.hot) {
+    module.hot.accept('./reducers', () => {
+      const reducers = require('./reducers').default
+      store.replaceReducer(reducers)
+    })
+  }
+
+  return store
+}
