+++ manoseimas/compatibility_test/client/store/reducers.js
@@ -9,9 +9,4 @@ export const makeRootReducer = (asyncReducers) => {
   })
 }
 
-export const injectReducer = (store, { key, reducer }) => {
-  store.asyncReducers[key] = reducer
-  store.replaceReducer(makeRootReducer(store.asyncReducers))
-}
-
 export default makeRootReducer
