+++ package.json
@@ -40,6 +40,7 @@
     "react-router": "^2.4.0",
     "react-router-redux": "^4.0.4",
     "redux": "^3.5.2",
+    "redux-thunk": "~2.1.0",
     "style-loader": "^0.13.1",
     "webpack": "^1.13.0",
     "webpack-bundle-tracker": "0.0.93",
